+++ src/main/common/README.md
@@ -1,1 +1,1 @@
-This folder contains common clients and repositories, and utility classes.
+This folder contains common clients and and utility classes.
