+++ src/main/reads/endpoints/UserActivityStatsEndpoint.ts
