+++ src/main/slack/App.ts
@@ -1,8 +1,8 @@
 // Clients
 import { HttpClient } from "../common/clients/HttpClient";
 import { SlackClient } from "./clients/SlackClient";
-import { IdentityRepository } from "../common/repositories/IdentityRepository";
-import { IdentityService } from "../common/services/IdentityService";
+import { IdentityRepository } from "../identity/repositories/IdentityRepository";
+import { IdentityService } from "../identity/services/IdentityService";
 import { OAuthService } from "./services/OAuthService";
 import { OAuthEndpoint } from "./endpoints/OAuthEndpoint";
 import { IdentityEndpoint } from "./endpoints/IdentityEndpoint"
